mally/requirements.txt
# Mally Multi-Agent System Requirements

# Core
asyncio>=3.4.3
numpy>=1.21.0
pyyaml>=6.0
dataclasses>=0.6;python_version<"3.7"

# Perception - YOLO/ONNX
onnxruntime-gpu>=1.15.0;platform_machine=="x86_64"
onnxruntime>=1.15.0;platform_machine!="x86_64"
opencv-python>=4.8.0
pillow>=9.0.0

# Navigation - SLAM
# ORB-SLAM3 requires manual installation
# scipy for optimization
scipy>=1.9.0

# Planning - MPC
casadi>=3.6.0
# acados-python  # Optional, requires separate installation

# Communication - MAVLink
pymavlink>=2.4.30
pyserial>=3.5

# Utilities
websockets>=11.0
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
